pyproject.toml
[tool.poetry]
name = "GamerDB"
version = "1.0.0"
description = ""
authors = ["StroupBSlayen <29642143+stroupbslayen@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.6.0"

aiosqlite = "^0.16.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"
python-dotenv = "^0.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
